pyproject.toml
[tool.poetry]
name = "dogcat"
version = "0.1.0"
description = ""
authors = ["lillian tsai"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
playwright = "^1.49.1"
pytest-playwright = "^0.6.2"
pytest-bdd = "^8.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
